backend/requirements.txt
fastapi==0.115.14
openai==1.93.0
pydantic==2.11.7
python-dotenv==1.1.1
supabase==2.16.0
tenacity==9.1.2
crewai
langchain
notion-client
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
markdown
